HBSD: Ensure that alternative signal stacks point to a stack mapping
We want to limit what kinds of mappings can be used for signal stacks.
Doing so prevents exploit payloads from using any kind of mapping,
especially for stack pivoting.

This is only one part of a coming series of commits for SROP
mitigations. Further research and development needs to happen to fully
mitigate SROP.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		14-STABLE