HBSD: maintain secure file ownership and permissions
This change maintains secure file ownership and permissions through
the periodic utility. The file list is described by a mtree(8) specification,
and the mtree(8) utility is run at daily intervals. The mtree(8) utility
will detect files that don't match the specification, and adjust ownership
and permissions only when necessary.

The mtree(8) specification is only used by periodic for the time being,
but if necessary, it could be used in other contexts. It could also be
modified to suit the user's preferences as well.

Credit to lattera / shawn.webb for suggesting the use of mtree(8). It works
much better than the previous approach that has been abandoned in favor
of mtree(8)