HBSD: Integrate PaX SEGVGUARD with Capsicum
Capsicum violations can be indicative of potential security issues. When
the kern.trap_enotcap sysctl node is set to a non-zero value, it's
likely that the user cares about these violations.

As such, integrate PaX SEGVGUARD here as a method of preventing
bruteforce attacks on capabilities-enabled processes. SEGVGUARD was
initially intended to protect against ASLR bruteforce attacks, but its
usage can be made broader so as to provide more protections.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Inspired-by:	Discussions with alip from Syd Linux
MFC-to:		14-STABLE