HBSD: Resolve merge conflict
With commit 9c4f1497dae7832e2727682e9161ca9572e56dfe, FreeBSD moved the
object file suffix transformation rules into a separate file. As part of
resolving this merge conflict, we need to remove references to the
PIE-related objects (.pieo and similar).

As part of the 2013-2015 ASLR work, we came up with a different method
for building with PIC/PIE. As such, we don't need to use FreeBSD's .pieo
and related.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>