HBSD: Enforce TPE during mmap(PROT_EXEC)
When Trusted Path Execution (TPE) is enabled, we need to check the file
descriptor passed into mmap(2) (obviously only when an fd is passed)
when executable permission is requested.

FreeBSD's filesystem path caching mechanism is a best-effort basis.
Meaning, attempting to look up the directory containing the file may
fail. In the case that the cache lookup misses, TPE is not enforced at
mmap(PROT_EXEC) time. This is a known limitation.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Reported-by:	@apache2
MFC-to:		14-STABLE